Nutrition Chatbot/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
requests